@@ -1,55 +1,66 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
+//apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "30.0.2"
+	compileSdkVersion 33
 
-    defaultConfig {
-        applicationId "com.lai.whiteboard"
-        minSdkVersion 19
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
+	defaultConfig {
+		applicationId "com.lai.whiteboard"
+		minSdkVersion 21
+		//noinspection ExpiredTargetSdkVersion
+		targetSdkVersion 29
+		versionCode 1
+		versionName "1.0"
 
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+		testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
-        externalNativeBuild {
-            cmake {
-                cppFlags "-std=c++11"
-            }
-        }
-        ndk {
-            abiFilters 'armeabi-v7a'
-        }
-    }
+		externalNativeBuild {
+			cmake {
+				cppFlags "-std=c++11"
+			}
+		}
+		ndk {
+			//noinspection ChromeOsAbiSupport
+			abiFilters 'armeabi-v7a'
+		}
+	}
 
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
+	buildTypes {
+		release {
+			minifyEnabled false
+			proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+		}
+	}
+	buildFeatures {
+		viewBinding true
+	}
 
-    externalNativeBuild {
-        cmake {
-            path "src/main/cpp/CMakeLists.txt"
-            version "3.10.2"
-        }
-    }
+	externalNativeBuild {
+		cmake {
+			path "src/main/cpp/CMakeLists.txt"
+			version "3.10.2"
+		}
+	}
+	namespace 'com.lai.whiteboard'
+
+	kotlinOptions {
+		jvmTarget = '1.8'
+	}
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'androidx.core:core-ktx:1.3.2'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+	implementation fileTree(dir: 'libs', include: ['*.jar'])
+	implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+	implementation 'androidx.appcompat:appcompat:1.6.1'
+	implementation 'androidx.preference:preference-ktx:1.2.0'
 
-    implementation "androidx.recyclerview:recyclerview:1.1.0"
-    implementation 'com.github.CymChad:BaseRecyclerViewAdapterHelper:3.0.4'
+	implementation 'androidx.core:core-ktx:1.10.0'
+	implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+	testImplementation 'junit:junit:4.13.2'
+	androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+	androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+
+	implementation "androidx.recyclerview:recyclerview:1.2.1"
+	implementation 'com.github.CymChad:BaseRecyclerViewAdapterHelper:3.0.4'
 }
