@@ -1,20 +1,25 @@
-package com.lai.whiteboard
+package com.lai.whiteboard.activities
 
 import android.Manifest
 import android.content.Intent
 import android.os.Build
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.lai.whiteboard.pen.page.WhiteboardActivity
-import kotlinx.android.synthetic.main.activity_main.*
+import com.lai.whiteboard.databinding.ActivityMainBinding
+
+// import kotlinx.android.synthetic.main.activity_main.*
 
 
 class MainActivity : AppCompatActivity() {
 
+    private lateinit var binding: ActivityMainBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        sample_text.setOnClickListener {
+        // setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        binding.sampleText.setOnClickListener {
             startActivity(Intent(this, WhiteboardActivity::class.java))
         }
 
