@@ -1,4 +1,4 @@
-package com.lai.whiteboard
+package com.lai.whiteboard.activities
 
 import android.graphics.BitmapFactory
 import android.opengl.GLSurfaceView
@@ -8,8 +8,11 @@ import android.util.Log
 import android.view.MotionEvent
 import android.widget.SeekBar
 import androidx.appcompat.app.AppCompatActivity
+import com.lai.whiteboard.R
+import com.lai.whiteboard.ShaderNative
+import com.lai.whiteboard.databinding.ActivityTestTextureBinding
 import com.lai.whiteboard.util.PointUtils
-import kotlinx.android.synthetic.main.activity_test_texture.*
+// import kotlinx.android.synthetic.main.activity_test_texture.*
 import javax.microedition.khronos.egl.EGLConfig
 import javax.microedition.khronos.opengles.GL10
 
@@ -17,7 +20,7 @@ import javax.microedition.khronos.opengles.GL10
 class TestTextureActivity : AppCompatActivity() {
 
     val bitmaps by lazy {
-        BitmapFactory.decodeResource(this.resources,R.mipmap.ic_qxx)
+        BitmapFactory.decodeResource(this.resources, R.mipmap.ic_qxx)
     }
 
     enum class DRAW_MODEL{
@@ -30,26 +33,29 @@ class TestTextureActivity : AppCompatActivity() {
     private var look_y:Float =0f
     private var look_z:Float =0f
 
+    private lateinit var binding: ActivityTestTextureBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_test_texture)
+        // setContentView(R.layout.activity_test_texture)
+        binding = ActivityTestTextureBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
-        gl_surface.setEGLContextClientVersion(3)
+        binding.glSurface.setEGLContextClientVersion(3)
 
 
-        gl_surface.setRenderer(object : GLSurfaceView.Renderer {
+        binding.glSurface.setRenderer(object : GLSurfaceView.Renderer {
             override fun onDrawFrame(gl: GL10?) {
-                if(currDraw==DRAW_MODEL.INIT){
+                if(currDraw== DRAW_MODEL.INIT){
                     ShaderNative.glTestDraw()
                 }else{
                     Log.e("!1111"," x $look_x y $look_y z $look_z")
-                    ShaderNative.glTestDrawR(look_x,look_y,look_z)
+                    ShaderNative.glTestDrawR(look_x, look_y, look_z)
                 }
             }
 
             override fun onSurfaceChanged(gl: GL10?, width: Int, height: Int) {
-                ShaderNative.glInit(width,height,bitmaps)
+                ShaderNative.glInit(width, height, bitmaps)
             }
 
             override fun onSurfaceCreated(gl: GL10?, config: EGLConfig?) {
@@ -58,11 +64,11 @@ class TestTextureActivity : AppCompatActivity() {
         })
 
 
-        x_seekBar.setOnSeekBarChangeListener(object :SeekBar.OnSeekBarChangeListener{
+        binding.xSeekBar.setOnSeekBarChangeListener(object :SeekBar.OnSeekBarChangeListener{
             override fun onProgressChanged(seekBar: SeekBar?, progress: Int, fromUser: Boolean) {
                 currDraw = DRAW_MODEL.DRAW_R
                 look_x = progress.toFloat()
-                gl_surface.requestRender()
+                binding.glSurface.requestRender()
 
             }
 
@@ -72,11 +78,11 @@ class TestTextureActivity : AppCompatActivity() {
             override fun onStopTrackingTouch(seekBar: SeekBar?) {
             }
         })
-        y_seekBar.setOnSeekBarChangeListener(object :SeekBar.OnSeekBarChangeListener{
+        binding.ySeekBar.setOnSeekBarChangeListener(object :SeekBar.OnSeekBarChangeListener{
             override fun onProgressChanged(seekBar: SeekBar?, progress: Int, fromUser: Boolean) {
                 currDraw = DRAW_MODEL.DRAW_R
                 look_y= progress.toFloat()
-                gl_surface.requestRender()
+                binding.glSurface.requestRender()
             }
 
             override fun onStartTrackingTouch(seekBar: SeekBar?) {
@@ -85,11 +91,11 @@ class TestTextureActivity : AppCompatActivity() {
             override fun onStopTrackingTouch(seekBar: SeekBar?) {
             }
         })
-        z_seekBar.setOnSeekBarChangeListener(object :SeekBar.OnSeekBarChangeListener{
+        binding.zSeekBar.setOnSeekBarChangeListener(object :SeekBar.OnSeekBarChangeListener{
             override fun onProgressChanged(seekBar: SeekBar?, progress: Int, fromUser: Boolean) {
                 currDraw = DRAW_MODEL.DRAW_R
                 look_z =progress.toFloat()
-                gl_surface.requestRender()
+                binding.glSurface.requestRender()
             }
 
             override fun onStartTrackingTouch(seekBar: SeekBar?) {
@@ -100,9 +106,9 @@ class TestTextureActivity : AppCompatActivity() {
         })
 
 
-        gl_surface.renderMode = RENDERMODE_WHEN_DIRTY
+        binding.glSurface.renderMode = RENDERMODE_WHEN_DIRTY
 
-        gl_surface.setOnTouchListener { v, event ->
+        binding.glSurface.setOnTouchListener { v, event ->
             when(event?.action){
                 MotionEvent.ACTION_DOWN->{
 
@@ -111,13 +117,15 @@ class TestTextureActivity : AppCompatActivity() {
                     val vertexWithPoint = PointUtils.vertexWithPoint(
                         event.x,
                         event.y,
-                        gl_surface.width,
-                        gl_surface.height
+                        binding.glSurface.width,
+                        binding.glSurface.height
                     )
                     look_x=vertexWithPoint[0]
                     look_y=vertexWithPoint[1]
                     currDraw = DRAW_MODEL.DRAW_R
-                    gl_surface.requestRender()
+                    binding.glSurface
+
+                        .requestRender()
                    // ShaderNative.glTestDrawR(vertexWithPoint[0],vertexWithPoint[1],1f)
                 }
             }
